Fix Build.PL to work with system-wide installs.
Previously circleator and the module that reads the Brewer
color specifications used FindBin to attempt to locate the
Perl library directory (../lib) and configuration file
directory (../conf). However, in a standard system-wide
install as root these two directories are not necessarily
located in the same place as the bin/ directory. The config
files go into /etc/circleator/conf and the library files
(on Ubuntu 13 at least) will go into /usr/local/share/perl.

To address this the following steps were taken:

1. A $DEFAULT_CONF_DIR global was added to the circleator Perl
script, as was a --conf_dir command-line option. If the option
is not specified then the default is used.
2. A $DEFAULT_LIB_DIR global was also added, along with a 'use
lib' statement to look for library modules in this directory.
3. Build.PL was modified to insert hard-coded values for these
two variables into the circleator Perl script before installing
it. Currently it uses a command-line Perl call to do the
replacement, which is not the most portable approach.

[Issue #12]